The home hero says the reason, not the date
The first thing anyone sees was five lines that fought each other.

The eyebrow read "THURSDAY, SEPTEMBER 3 · TONIGHT" — thirty characters
of tracked caps, with the proof the page is fresh set in front of the
one word that says why this game is on the screen. With a publisher's
mark on show it grew a third clause and read "THURSDAY, SEPTEMBER 3 ·
TONIGHT · CONTINUE". The date is now its own field on the slide, set
after the reason and in a dimmer grey, and abbreviated: "TONIGHT
Thu, 3 Sep".

The headline was a sentence with a verb in it — "Continue GreedFall:
The Dying World" — over a button reading "Continue it" two hundred
points below. The largest type on the page spent a word repeating the
smallest, and cost the title a line doing it. It is the game's name
now, on every slide, which is also what lets the publisher's mark stand
in for the line everywhere rather than only where a verb could be
pushed out of the way.

The one number the reader came for was buried mid-sentence in grey body
copy: "2.5h left. Already under way — chip away at it." Every other
figure in this app is set as a figure. "2.5h left" is now amber label
type with the sentence after it, and each slide's sentence stands on
its own so it reads either way.

And the stage knew how far through the game was and drew nothing. It is
a bar under the title now, amber on white, absent — not empty — for a
game nobody has started, because a bar at zero in the loudest place on
the page says "you have not begun this" about the game being
recommended. The fraction comes off the library, where it always was.

Two smaller things: the copy block sits at thirty-two off the bottom
rather than twenty, so the action row is inside the picture instead of
on the last rows of the stage where the artwork has already faded to
the page's grey — the desk got that fix and the phone did not. And the
page dots go from five points at a third opacity to six at a half:
over a photograph they read as dust, and they were the only mark on
the stage saying there was more than one slide.

One test changed for a reason worth recording. `index.test.tsx` claimed
a saved game must not appear on the home screen, and asserted it
against the whole screen; it passed only because the stage's headline
was "Start Game 3" rather than "Game 3". The moment the verb moved to
the button it failed on correct behaviour — the shortest thing you
saved is exactly what the Tonight slide is for. It now asserts the game
appears once, at the top, and nowhere in the shelves.

+++ src/hooks/useStage.ts
@@ -79,11 +79,17 @@ export function useStage({
             short,
             trending,
             outToday: count ?? 0,
+            /**
+             * Abbreviated. Spelled out, "Thursday, September 3" was
+             * twenty-one characters of tracked caps standing in front
+             * of the word that says why the game below it is there.
+             * The date is proof the page is today's, not a headline.
+             */
             dateLabel: hydrated
               ? new Date(now).toLocaleDateString(undefined, {
-                  weekday: 'long',
+                  weekday: 'short',
                   day: 'numeric',
-                  month: 'long',
+                  month: 'short',
                 })
               : '',
           }),
